Update Yu-Gi-Oh.py
set up the initial battle phase where each opponent draws five cards from their decks and every card is random. and each of the opponent starts out with 8000 life points. and on top of that you can click on each of the cards individually and you can see each of the cards enlarged.