Update Yu-Gi-Oh.py
When the user clicks on the info button it will display the information of the card such as the name and full description of the card. and it will also properly display the attack, defense, and race of the card if it is a monster. on the other hand if it is a spell or trap it will display the type of card it is, such as normal or continuous.