Update CardEffects.py
Separated the choice of the user from the change of the field  into two functions, command(User's choice) and MainCommand(how the field changes based on the user's choice/AI). this way when the cards effects are implemented it will work for the user and the AI. the function Decoder() will expand the array by providing the necessary card information for the MainCommand() to execute. this way the AI will just have simplified output that it needs to compute.