Update CardEffects.py
the function SummonSetCard() Informs which button can be activated and which cannot based on the level of the monster and the number of cards that is on the field.  Inside the Command() function I have compacted both of the wile loops into a single while loop since the check is being done with the SummonSetCard() function. On top of that the Save function saves self.fieldInfo (which contains everything that the current player knows) and command(what the player wants to do) into the database for the training process of deep learning.