Update CardEffects.py
I have tinker with how the information with in dataFieldInfo with in the function convert and the class variable self.fieldInfo handles the data. I did this so that the information is more subdivided. This way the deep learning can properly evaluate the field just by what is in the database by a single string. and the output has all been given a constant length so that it can make the training process much easier to manage.